피드백: 리팩토링
