+++ src/components/home/Friends.tsx
