+++ src/components/home/GiftInfo.tsx
