+++ src/pages/ThemePage.tsx
@@ -1,33 +1,22 @@
 import { useEffect, useRef, useState, useCallback } from "react";
-import { useParams, useNavigate } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import { getThemeProduct } from "@/services/theme";
 import styled from "@emotion/styled";
 import axios from "axios";
 import { showErrorToast } from "@/styles/toast";
-import Spacing from "@/components/Spacing";
 import type { ProductInfo } from "@/types/product";
-import { getUserFromSession } from "@/utils/getUserFromStorage";
-import { PATH } from "@/paths";
-import { useFetchTheme } from "./useFetchTheme";
+import { useFetchTheme } from "../hooks/useFetchTheme";
+import ThemeHeader from "@/components/theme/ThemeHeader";
+import ProductGrid from "@/components/theme/ProductGrid";
 
 export default function ThemePage() {
   const { themeId } = useParams();
-  const navigate = useNavigate();
   const { theme, loading: themeLoading } = useFetchTheme(themeId);
   const [products, setProducts] = useState<ProductInfo[]>([]);
   const [loading, setLoading] = useState(false);
   const [hasMore, setHasMore] = useState(true); // 더 불러 올 게 있는지
   const pageRef = useRef(1); // 현재 페이지 번호 저장
   const loader = useRef<HTMLDivElement | null>(null); // 마지막 감지
-
-  const goToOrder = (itemId: number) => {
-    const userInfo = getUserFromSession();
-    if (userInfo) navigate(PATH.toORDER(itemId));
-    else
-      navigate(PATH.LOGIN, {
-        state: { from: `/order/${itemId}` },
-      });
-  };
 
   // 상품 불러오기
   const fetchProducts = useCallback(
@@ -82,40 +71,9 @@ export default function ThemePage() {
 
   return (
     <Wrapper>
-      <HeaderWrapper backgroundColor={theme.backgroundColor}>
-        <Name>{theme.name}</Name>
-        <Spacing height="8px" />
-        <Title>{theme.title}</Title>
-        <Spacing height="4px" />
-        <Description>{theme.description}</Description>
-      </HeaderWrapper>
-      <CardWrapper>
-        {products.length === 0 && loading ? (
-          <Spinner />
-        ) : products.length === 0 ? (
-          <EmptyBox>
-            <EmptyMessage>상품이 없습니다.</EmptyMessage>
-          </EmptyBox>
-        ) : (
-          <CardGrid>
-            {products.map((item) => (
-              <Card key={item.id} onClick={() => goToOrder(item.id)}>
-                <Image src={item.imageURL} />
-                <Spacing height="12px" />
-                <Brand>{item.brandInfo.name}</Brand>
-                <Label>{item.name}</Label>
-                <Spacing height="4px" />
-                <Price>
-                  {item.price.sellingPrice.toLocaleString()}
-                  <PriceUnit>원</PriceUnit>
-                </Price>
-              </Card>
-            ))}
-            <div ref={loader} style={{ height: 1 }} />
-          </CardGrid>
-        )}
-        {loading && <Spinner />}
-      </CardWrapper>
+      <ThemeHeader theme={theme} />
+      <ProductGrid products={products} loader={loader} loading={loading} />
+      {loading && <Spinner />}
     </Wrapper>
   );
 }
@@ -123,102 +81,6 @@ export default function ThemePage() {
 const Wrapper = styled.div`
   width: 100%;
   height: 100%;
-`;
-
-const HeaderWrapper = styled.section<{ backgroundColor: string }>`
-  width: 100%;
-  padding: 1.625rem 1rem 1.375rem;
-  background-color: ${({ backgroundColor }) => backgroundColor};
-`;
-
-const Name = styled.p`
-  ${({ theme }) => theme.typography.label1Bold};
-  color: ${({ theme }) => theme.colors.gray[100]};
-  margin: 0px;
-  text-align: left;
-`;
-
-const Title = styled.h5`
-  ${({ theme }) => theme.typography.title1Bold};
-  color: ${({ theme }) => theme.colors.gray.white};
-  margin: 0px;
-  text-align: left;
-`;
-
-const Description = styled.p`
-  ${({ theme }) => theme.typography.body1Regular};
-  color: ${({ theme }) => theme.colors.gray[200]};
-  margin: 0px;
-  text-align: left;
-`;
-
-const CardWrapper = styled.div`
-  padding: 16px;
-  width: 100%;
-`;
-
-const CardGrid = styled.div`
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 24px 8px;
-  align-itmes: stretch;
-`;
-
-const Card = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  background-color: transparent;
-  cursor: pointer;
-  width: 100%;
-  height: 100%;
-`;
-
-const Image = styled.img`
-  width: 100%;
-  aspect-ratio: 1 / 1;
-  object-fit: cover;
-  border-radius: 4px;
-  background-color: ${({ theme }) => theme.colors.gray[200]};
-`;
-
-const Brand = styled.p`
-  ${({ theme }) => theme.typography.body2Regular};
-  color: ${({ theme }) => theme.colors.gray[600]};
-  margin: 0px;
-  text-align: left;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
-
-const Label = styled.h6`
-  ${({ theme }) => theme.typography.body2Regular};
-  color: ${({ theme }) => theme.colors.gray[900]};
-  margin: 0px;
-  text-align: left;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-`;
-
-const Price = styled.p`
-  ${({ theme }) => theme.typography.title2Bold};
-  color: ${({ theme }) => theme.colors.gray[900]};
-  margin: 0px;
-  text-align: left;
-  word-break: break-word;
-`;
-
-const PriceUnit = styled.span`
-  ${({ theme }) => theme.typography.subtitle1Regular};
-  color: ${({ theme }) => theme.colors.gray[900]};
-  margin: 0px;
-  text-align: left;
-  word-break: break-word;
 `;
 
 const Spinner = styled.div`
@@ -235,22 +97,4 @@ const Spinner = styled.div`
       transform: rotate(360deg);
     }
   }
-`;
-
-const EmptyBox = styled.div`
-  width: 100%;
-  height: 240px;
-  display: flex;
-  -webkit-box-pack: center;
-  justify-content: center;
-  -webkit-box-align: center;
-  align-items: center;
-`;
-
-const EmptyMessage = styled.p`
-  ${({ theme }) => theme.typography.body2Regular};
-  color: ${({ theme }) => theme.colors.gray[900]};
-  margin: 0px;
-  width: 100%;
-  text-align: center;
 `;
