+++ src/types/theme.ts
@@ -1,0 +1,8 @@
+export type ThemeInfo = {
+  themeId: number;
+  name: string;
+  image: string;
+  title: string;
+  description: string;
+  backgroundColor: string;
+};
